fixed lots of execution bugs, added opcodes, more
Fixed many execution bugs. My method of comparing output with other
emulators was apparantly just 100% borked, leading me to believe that my
emulator was consistent with others. It was not. It still isn't, but
it's closer, and I'm not certain that it's necessarily incorrect.

Blargg's cpu tests now properly print the ascii sent to the link port,
so that test seems to actually be running correctly. The data in VRAM
also seems to be _mostly_ right. Dumping VRAM and copying it into the
BGB emulator works successfully, and all of the necessary tiles are
properly stored.