a start to opcode gen. currently NOP, LD, INC
This branch will be the main working branch until it becomes at least as
functional as the master branch. I have moved to izik's table for
codegen rather than the pastraiser table due to mistakes in the
pastraiser table. I figure this is also a good time to update the
codegen to actually work well, rather than the current solution I have
that's full of 1-off hacks. Hopefully this works out well in the long
run.