big opcode restructure: moved opcodes to new file
Overall, this is a slight performence _loss_. Although I no longer have
the huge if/elsif chain (which case statements expand to), the overhead
of function calls still deecreases performance.

To measure this, I counted the time leftover after each frame for 1000
frames, running that trial 5-6 times for both implementations. With the
case-statement approach, there was an average of 9.2s spent waiting
after frames. With the function-call approach, there was an average of
8.8s spent waiting after frames. This means that each batch of 70224
cycles now takes about 0.0004s longer than it would have before.

However, I believe that this is a good change purely for legibility.