removed APU due to Crystal GC issues..............
There's a comment in APu.cr with more details. Who knows if this will be
fixed. I've spent over a day debugging and soliciting help in the
Crystal gitter channel. At this point, I'm going to focus on fixing up
other parts of the emulator. I've added bindings for nanosleep, which
seems to yield pretty accurate timing now that audio sync is gone... :/