an actual step forwards in audio processing (audio sync, channel 2, less bugs)
I've discovered that stack-allocated arrays aren't subject to the same
GC issue that heap-allocated arrays/slices are, meaning I can use them
to store buffers that I pass to SDL. The only problem with StaticArrays
is that there's another bug in the Crystal compiler that leads to
_extremely_ long compile times when compiling StaticArrays for release.
However, some people in the gitter channel have figured out a
workaround, which I've added to util.cr. There's an open issue and a PR
on GitHub that address this problem, so hopefully they'll be reviewed
soon.

Given these changes, I've moved back to audio syncing, and now 2
channels are implemented and abstracted (somewhat). Audio is now
actually a feasible option.