replace cgb func w/ bool ptr, add bg tile rotation
The Motherboard.cgb function was taking a bit too long, making the audio
sound worse when not compiling for release. This is because it was
effectively calling an additional 92,120 functions per frame (over 5.5m
function calls per second). Now, I just pass a pointer which I update
when the bootrom gets unmapped.

Also added bg and window tile rotation, both vertical and horizontal.