major refactor, improve scheduler efficiency
=== Major refactor
All components are now passed a copy of the Motherboard itself, and can
access components directly off of that. This removes circular
dependencies between components, which will _hopefully_ allow for some
cleaner abstractions in the future.

=== Scheduler efficiency
Events are now represented as structs because the GC was eating up tons
of time. Additionally, cycles are now counted in the scheduler to avoid
mutation of scheduled events on each cycle.