put apu on scheduler, more details in detail message
This required changing around some of the hacky overrides on the wave
channel. Where I used to stub 4 cycles, I now need to stub 6. This still
passes blargg's cgb_sound tests, but it'd be nice to find the root cause
of these stubbed cycles in the first place, as well as why they've just
changed to requiring 6. My first theory is that the channels aren't
turning off as they should be any more, since I don't clear the
scheduled events when the channels/APU are turned off. That's likely a
big `todo` in the future for better accuracy, but I'm going to leave it
for now for the performance improvements.